+++ plugins/shortcuts/back.js
@@ -1,9 +1,8 @@
-const { globalShortcut, ipcMain } = require("electron");
+const { globalShortcut } = require("electron");
 const is = require("electron-is");
 const electronLocalshortcut = require("electron-localshortcut");
 const getSongControls = require("../../providers/song-controls");
-const { setupMPRIS } = require("./mpris");
-const registerCallback = require("../../providers/song-info");
+const registerMPRIS = require("./mpris");
 
 function _registerGlobalShortcut(webContents, shortcut, action) {
 	globalShortcut.register(shortcut, () => {
@@ -30,7 +29,7 @@ function registerShortcuts(win, options) {
 	_registerLocalShortcut(win, "CommandOrControl+F", search);
 	_registerLocalShortcut(win, "CommandOrControl+L", search);
 
-	if (is.linux()) registerMPRIS();
+	if (is.linux()) registerMPRIS(win);
 
 	const { global, local } = options;
 	const shortcutOptions = { global, local };
@@ -56,72 +55,6 @@ function registerShortcuts(win, options) {
 			} else { // type === "local"
 				_registerLocalShortcut(win, local[action], songControls[action]);
 			}
-		}
-	}
-	function registerMPRIS() {
-		try {
-			const secToMicro = n => Math.round(Number(n) * (1000 * 1000));
-			const microToSec = n => Math.round(Number(n) / 1000 / 1000);
-
-			const seekTo = e => win.webContents.send("seekTo", microToSec(e.position));
-			const seek = o => win.webContents.send("seek", microToSec(o));
-
-			const player = setupMPRIS();
-
-			const mprisSeek = player.seeked;
-
-			win.webContents.send("registerOnSeek");
-
-			ipcMain.on('seeked', (_, t) => mprisSeek(secToMicro(t)));
-
-			let currentSeconds = 0;
-			ipcMain.on('timeChanged', (_, t) => currentSeconds = t);
-
-			player.getPosition = () => secToMicro(currentSeconds)
-
-			player.on("raise", () => {
-				win.setSkipTaskbar(false);
-				win.show();
-			});
-
-			player.on("play", () => {
-				if (player.playbackStatus !== 'Playing') {
-					player.playbackStatus = 'Playing';
-					playPause()
-				}
-			});
-			player.on("pause", () => {
-				if (player.playbackStatus !== 'Paused') {
-					player.playbackStatus = 'Paused';
-					playPause()
-				}
-			});
-
-			player.on("playpause", playPause);
-			player.on("next", next);
-			player.on("previous", previous);
-
-			player.on('seek', seek);
-			player.on('position', seekTo);
-
-			registerCallback(songInfo => {
-				if (player) {
-					const data = {
-						'mpris:length': secToMicro(songInfo.songDuration),
-						'mpris:artUrl': songInfo.imageSrc,
-						'xesam:title': songInfo.title,
-						'xesam:artist': songInfo.artist,
-						'mpris:trackid': '/'
-					};
-					if (songInfo.album) data['xesam:album'] = songInfo.album;
-					player.metadata = data;
-					mprisSeek(secToMicro(songInfo.elapsedSeconds))
-					player.playbackStatus = songInfo.isPaused ? "Paused" : "Playing"
-				}
-			})
-
-		} catch (e) {
-			console.warn("Error in MPRIS", e);
 		}
 	}
 }
