+++ plugins/shortcuts/mpris.js
@@ -1,4 +1,6 @@
 const mpris = require("mpris-service");
+const { ipcMain } = require("electron");
+const registerCallback = require("../../providers/song-info");
 
 function setupMPRIS() {
 	const player = mpris({
@@ -14,6 +16,71 @@ function setupMPRIS() {
 	return player;
 }
 
-module.exports = {
-	setupMPRIS,
-};
+function registerMPRIS(win) {
+	try {
+		const secToMicro = n => Math.round(Number(n) * 1e6);
+		const microToSec = n => Math.round(Number(n) / 1e6);
+
+		const seekTo = e => win.webContents.send("seekTo", microToSec(e.position));
+		const seek = o => win.webContents.send("seek", microToSec(o));
+
+		const player = setupMPRIS();
+
+		const mprisSeek = player.seeked;
+
+		win.webContents.send("registerOnSeek");
+
+		ipcMain.on('seeked', (_, t) => mprisSeek(secToMicro(t)));
+
+		let currentSeconds = 0;
+		ipcMain.on('timeChanged', (_, t) => currentSeconds = t);
+
+		player.getPosition = () => secToMicro(currentSeconds)
+
+		player.on("raise", () => {
+			win.setSkipTaskbar(false);
+			win.show();
+		});
+
+		player.on("play", () => {
+			if (player.playbackStatus !== 'Playing') {
+				player.playbackStatus = 'Playing';
+				playPause()
+			}
+		});
+		player.on("pause", () => {
+			if (player.playbackStatus !== 'Paused') {
+				player.playbackStatus = 'Paused';
+				playPause()
+			}
+		});
+
+		player.on("playpause", playPause);
+		player.on("next", next);
+		player.on("previous", previous);
+
+		player.on('seek', seek);
+		player.on('position', seekTo);
+
+		registerCallback(songInfo => {
+			if (player) {
+				const data = {
+					'mpris:length': secToMicro(songInfo.songDuration),
+					'mpris:artUrl': songInfo.imageSrc,
+					'xesam:title': songInfo.title,
+					'xesam:artist': songInfo.artist,
+					'mpris:trackid': '/'
+				};
+				if (songInfo.album) data['xesam:album'] = songInfo.album;
+				player.metadata = data;
+				mprisSeek(secToMicro(songInfo.elapsedSeconds))
+				player.playbackStatus = songInfo.isPaused ? "Paused" : "Playing"
+			}
+		})
+
+	} catch (e) {
+		console.warn("Error in MPRIS", e);
+	}
+}
+
+module.exports = registerMPRIS;
